refactor liquidity tests
